feat: add toggle between different signature flavours
The specification of JSON Web Signature 2020 is not specific about how the payload is assembled. As a consequence, services related to "Gaia-X" have implemented a suite that differs from other implementations. The spec provides a list of test vectors to validate implementations. This adds an implementation that works with the test vectors and adds unit tests to validate this behaviour.